requirements.txt
aiodns==2.0.0
aiofiles==0.4.0
aiohttp==4.0.0a1
brotlipy==0.7.0
chardet==3.0.4
efficientnet==1.1.0
graphviz==0.13.2
h5py==2.10.0
joblib==0.14.1
jupyter==1.0.0
matplotlib==3.2.0
numba
numpy
pandas>=1.0.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pydot-ng==2.0.0
requests==2.23.0
scikit-learn
scipy==1.4.1
tenacity==6.1.0
tensorboard==2.2.0
tensorflow==2.2.0rc2
tqdm==4.43.0
urllib3==1.25.8
tensorboard_plugin_profile
tensorflow-addons
runnb
